// src/sites/pigu/types.ts
export type WigdetData = {
    productId: number
    m: number
    v: number
    h: string
    w: number
    csrf_p_token: string
    ua: string
    photoUrl: string
    title: string
    url: string
    wishlistId: number
    pl: number
    meta: Meta
    categoryId: string
    n18: string
    productClass: null
    supplierId: null
    specialMessage: number
}

export type Meta = {
    id: number
    title: string
    sell_price: number
    location: string
}